requirements.txt
Flask
gunicorn
openai
gspread
google-auth
google-auth-oauthlib
selenium
webdriver-manager
requests
python-dotenv
openai
flask
gspread
oauth2client
pydrive
python-dotenv
requests
beautifulsoup4
transformers
torch
pandas
streamlit
certifi
selenium
